Rename subcommand name from server to core
