+++ commands/core/example/cmd.go
