+++ commands/core/example/example.yaml
